+++ efinance/__init__.py
@@ -2,11 +2,6 @@
 .. include:: ../README.md
 """
 __docformat__ = "restructuredtext"
-from efinance.api import (
-    bond,
-    fund,
-    futures,
-    stock,
-)
+from efinance.api import bond, fund, futures, stock
 
 __all__ = ['stock', 'fund', 'bond', 'futures']